requirements.txt
injector
Flask==1.1.1
requests==2.22.0
Flask_Cors==3.0.8
Flask-Ext==0.1
Flask-Injector==0.12.3
flask-restplus==0.13.0
gast==0.3.3
Werkzeug==0.16.0
xlrd==1.2.0
jsonpickle==1.3
PyYAML==5.3.1
six==1.14.0
pyjwkest==1.4.2
pandas==0.25.1
numpy==1.16.5
matplotlib==3.1.1
scikit_learn==0.23.1
scipy==1.4.1
keras==2.3.1
tensorflow==2.2.0

